Last push
Fell asleep while working. As a result the project did not get finished
and the frontend was never integrated with the backend. The version we
will show to the judges has several hardcoded values as opposed to the
values from the backend. I might fix the issues on my own time but as
for HackGT this is as far as it goes.